+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryConfig.java
@@ -73,6 +73,8 @@ public interface BigQueryConfig {
 
   Optional<Long> getCreateReadSessionTimeoutInSeconds();
 
+  int getChannelPoolSize();
+
   // Get a static flow control window per RPC. When not set
   // auto flow control is determined by Bandwidth Delay Product.
   Optional<Integer> getFlowControlWindowBytes();